Intergrating the IGDB api into production.
Added a function for searching for games in the IGDB database. This search function generates a list of links to details of that particular game.

To do this I had to proxy the IGDB api through backend. This is because of how they use the access token in their api.

Fixed: Hid all keys/id/secrets in env variables.